Task refactor and user limits

+++ back/src/main/java/com/ionnote/services/TaskService.java
@@ -2,7 +2,9 @@ package com.ionnote.services;
 
 import com.ionnote.dtos.task.*;
 import com.ionnote.entities.Task;
+import com.ionnote.exceptions.ForbiddenException;
 import com.ionnote.exceptions.TaskNotFoundException;
+import com.ionnote.exceptions.UserNotFoundException;
 import com.ionnote.repositories.TaskRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.BeanUtils;
@@ -15,11 +17,15 @@ import java.util.UUID;
 @AllArgsConstructor
 public class TaskService {
     private TaskRepository taskRepository;
+    private AuthService authService;
 
     //Create
-    public void createTask(CreateTaskDTO dto) {
+    public void createTask(CreateTaskDTO dto) throws UserNotFoundException {
+        var loggedUser = authService.getLoggedUser();
+
         var tempTask = Task.builder()
-                .uuid(UUID.randomUUID())
+                .uuid(UUID.randomUUID().toString())
+                .ownerId(loggedUser.getId())
                 .name(dto.getName())
                 .deadline(dto.getDeadline())
                 .completed(Boolean.FALSE)
@@ -29,15 +35,20 @@ public class TaskService {
     }
 
     //Read
-    public TaskDTO readTask(ReadTaskDTO dto) throws TaskNotFoundException {
+    public TaskDTO readTask(ReadTaskDTO dto) throws TaskNotFoundException, UserNotFoundException, ForbiddenException {
         var task = taskRepository.findById(dto.getUuid()).orElseThrow(TaskNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(task.getOwnerId())){
+            throw new ForbiddenException();
+        }
         var response = new TaskDTO();
         BeanUtils.copyProperties(task,response);
         return response;
     }
 
-    public List<TaskDTO> readAllTasks() {
-        var tasks = taskRepository.findAll();
+    public List<TaskDTO> readAllTasks() throws UserNotFoundException {
+        var loggedUser = authService.getLoggedUser();
+        var tasks = taskRepository.findAllByOwnerId(loggedUser.getId());
         return tasks.stream().map(task -> {
             var tempResponse = new TaskDTO();
             BeanUtils.copyProperties(task,tempResponse);
@@ -46,15 +57,23 @@ public class TaskService {
     }
 
     //Update
-    public void updateTask(UpdateTaskDTO dto) throws TaskNotFoundException {
+    public void updateTask(UpdateTaskDTO dto) throws TaskNotFoundException, UserNotFoundException, ForbiddenException {
         var task = taskRepository.findById(dto.getUuid()).orElseThrow(TaskNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(task.getOwnerId())){
+            throw new ForbiddenException();
+        }
         BeanUtils.copyProperties(dto, task);
         taskRepository.save(task);
     }
 
     //Delete
-    public void deleteTask(DeleteTaskDTO dto) throws TaskNotFoundException {
+    public void deleteTask(DeleteTaskDTO dto) throws TaskNotFoundException, UserNotFoundException, ForbiddenException {
         var task = taskRepository.findById(dto.getUuid()).orElseThrow(TaskNotFoundException::new);
+        var loggedUser = authService.getLoggedUser();
+        if (!loggedUser.getId().equals(task.getOwnerId())){
+            throw new ForbiddenException();
+        }
         taskRepository.delete(task);
     }
